Refactor #1 - terms page
